Remove placeholder Firebase config and update imports
