--- lib/src/core/firebase_options.dart
@@ -1,60 +1,0 @@
-import 'package:firebase_core/firebase_core.dart' show FirebaseOptions;
-import 'package:flutter/foundation.dart' show defaultTargetPlatform, kIsWeb, TargetPlatform;
-
-/// Placeholder Firebase configuration.
-///
-/// Replace the values in this file with your actual Firebase project
-/// settings or regenerate using the FlutterFire CLI.
-class DefaultFirebaseOptions {
-  static FirebaseOptions get currentPlatform {
-    if (kIsWeb) {
-      return web;
-    }
-    switch (defaultTargetPlatform) {
-      case TargetPlatform.android:
-        return android;
-      case TargetPlatform.iOS:
-        return ios;
-      case TargetPlatform.macOS:
-        return macos;
-      default:
-        return web;
-    }
-  }
-
-  static const FirebaseOptions web = FirebaseOptions(
-    apiKey: 'REPLACE_WITH_API_KEY',
-    appId: 'REPLACE_WITH_APP_ID',
-    messagingSenderId: 'REPLACE_WITH_MESSAGING_SENDER_ID',
-    projectId: 'REPLACE_WITH_PROJECT_ID',
-    authDomain: 'REPLACE_WITH_AUTH_DOMAIN',
-    storageBucket: 'REPLACE_WITH_STORAGE_BUCKET',
-    measurementId: 'REPLACE_WITH_MEASUREMENT_ID',
-  );
-
-  static const FirebaseOptions android = FirebaseOptions(
-    apiKey: 'REPLACE_WITH_API_KEY',
-    appId: 'REPLACE_WITH_APP_ID',
-    messagingSenderId: 'REPLACE_WITH_MESSAGING_SENDER_ID',
-    projectId: 'REPLACE_WITH_PROJECT_ID',
-    storageBucket: 'REPLACE_WITH_STORAGE_BUCKET',
-  );
-
-  static const FirebaseOptions ios = FirebaseOptions(
-    apiKey: 'REPLACE_WITH_API_KEY',
-    appId: 'REPLACE_WITH_APP_ID',
-    messagingSenderId: 'REPLACE_WITH_MESSAGING_SENDER_ID',
-    projectId: 'REPLACE_WITH_PROJECT_ID',
-    iosBundleId: 'com.example.clearskyPhotoReports',
-    storageBucket: 'REPLACE_WITH_STORAGE_BUCKET',
-  );
-
-  static const FirebaseOptions macos = FirebaseOptions(
-    apiKey: 'REPLACE_WITH_API_KEY',
-    appId: 'REPLACE_WITH_APP_ID',
-    messagingSenderId: 'REPLACE_WITH_MESSAGING_SENDER_ID',
-    projectId: 'REPLACE_WITH_PROJECT_ID',
-    iosBundleId: 'com.example.clearskyPhotoReports',
-    storageBucket: 'REPLACE_WITH_STORAGE_BUCKET',
-  );
-}
